Remove some duplicate code
